Changes to completely remove all floating point ops if SQLITE_OMIT_FLOATING_POINT defined.  Note that w/o fp, date/time, round, nan, etc. are all gone or limited in functionality.   Updated some of the test scripts to support missing fp and 64-bit functionality.   Ticket #3029.

+++ src/vdbeapi.c
@@ -752,7 +752,8 @@ static Mem *columnMem(sqlite3_stmt *pStmt, int i){
     vals = sqlite3_data_count(pStmt);
     pOut = &pVm->pResultSet[i];
   }else{
-    static const Mem nullMem = {{0}, 0.0, 0, "", 0, MEM_Null, SQLITE_NULL, 0, 0, 0 };
+    /* ((double)0) In case of SQLITE_OMIT_FLOATING_POINT... */
+    static const Mem nullMem = {{0}, (double)0, 0, "", 0, MEM_Null, SQLITE_NULL, 0, 0, 0 };
     if( pVm->db ){
       sqlite3_mutex_enter(pVm->db->mutex);
       sqlite3Error(pVm->db, SQLITE_RANGE, 0);
